기존 API 경로 수정
